+++ CHANGELOG.md
@@ -10,18 +10,24 @@
 import { Toast } from 'buefy/dist/es/components/toast'
 Toast.open('Toasty!')
 ```
+* Fix #1529 pagination ellipsis behavior (thanks @service-paradis)
+* Fix #1550 improve the default behavior for clickable steps (thanks @service-paradis)
 
 ### New features
 
+* **New component: Progress Bar**, [see docs](https://buefy.org/documentation/progress)
 * Improve to tree shaking when importing individual components
 * Add ``defaultTooltipDelay`` constructor option (thanks @service-paradis)
-* Add ``sort-icon`` and ``sort-icon-size`` to table (@modularcoder)
+* Add ``sort-icon`` and ``sort-icon-size`` props to table (@modularcoder)
 * Add ``default-minutes`` and ``default-seconds`` to timepicker
+* Add ``focusable`` prop to dropdown (thanks @service-paradis)
+* Add ``mouseenter `` and ``mouseleave`` events to table (thanks @afm-sayem)
 
 ### Fixes
 
 * Fix #1541 emit `input` event when change step (thanks @service-paradis)
 * Fix #1544 years select datepicker when max/min being passed
+* Fix issues on delete last selected element and on reset upload (thanks @it-fm)
 
 ## 0.7.10